+++ src/lib/permissions.ps1
@@ -1,5 +1,34 @@
 <#
 .SYNOPSIS
+    Set the custom Cosmos DB User types to the permission
+    returned by an API call.
+
+.DESCRIPTION
+    This function applies the custom types to the permission returned
+    by an API call.
+
+.PARAMETER Permission
+    This is the permission that is returned by a user API call.
+#>
+function Set-CosmosDbPermissionType
+{
+    [CmdletBinding()]
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        $Permission
+    )
+
+    foreach ($item in $Permission)
+    {
+        $item.PSObject.TypeNames.Insert(0, 'CosmosDB.Permission')
+    }
+
+    return $Permission
+}
+
+<#
+.SYNOPSIS
     Return the resource path for a permission object.
 
 .DESCRIPTION
@@ -80,7 +109,7 @@ function Get-CosmosDbPermission
     (
         [Parameter(Mandatory = $true, ParameterSetName = 'Connection')]
         [ValidateNotNullOrEmpty()]
-        [PSCustomObject]
+        [CosmosDb.Connection]
         $Connection,
 
         [Parameter(Mandatory = $true, ParameterSetName = 'Account')]
@@ -122,17 +151,24 @@ function Get-CosmosDbPermission
     {
         $null = $PSBoundParameters.Remove('Id')
 
-        return Invoke-CosmosDbRequest @PSBoundParameters `
+        $permission = Invoke-CosmosDbRequest @PSBoundParameters `
             -Method 'Get' `
             -ResourceType 'permissions' `
             -ResourcePath ('{0}/{1}' -f $resourcePath, $Id)
     }
     else
     {
-        return Invoke-CosmosDbRequest @PSBoundParameters `
+        $result = Invoke-CosmosDbRequest @PSBoundParameters `
             -Method 'Get' `
             -ResourceType 'permissions' `
             -ResourcePath $resourcePath
+
+        $permission = $result.Permissions
+    }
+
+    if ($permission)
+    {
+        return (Set-CosmosDbPermissionType -Permission $permission)
     }
 }
 
@@ -181,7 +217,7 @@ function New-CosmosDbPermission
     (
         [Parameter(Mandatory = $true, ParameterSetName = 'Connection')]
         [ValidateNotNullOrEmpty()]
-        [PSCustomObject]
+        [CosmosDb.Connection]
         $Connection,
 
         [Parameter(Mandatory = $true, ParameterSetName = 'Account')]
@@ -232,11 +268,16 @@ function New-CosmosDbPermission
 
     $resourcePath = ('users/{0}/permissions' -f $UserId)
 
-    return Invoke-CosmosDbRequest @PSBoundParameters `
+    $permission = Invoke-CosmosDbRequest @PSBoundParameters `
         -Method 'Post' `
         -ResourceType 'permissions' `
         -ResourcePath $resourcePath `
         -Body "{ `"id`": `"$id`", `"permissionMode`" : `"$PermissionMode`", `"resource`" : `"$Resource`" }"
+
+    if ($permission)
+    {
+        return (Set-CosmosDbPermissionType -Permission $permission)
+    }
 }
 
 <#
@@ -276,7 +317,7 @@ function Remove-CosmosDbPermission
     (
         [Parameter(Mandatory = $true, ParameterSetName = 'Connection')]
         [ValidateNotNullOrEmpty()]
-        [PSCustomObject]
+        [CosmosDb.Connection]
         $Connection,
 
         [Parameter(Mandatory = $true, ParameterSetName = 'Account')]
@@ -315,7 +356,7 @@ function Remove-CosmosDbPermission
 
     $resourcePath = ('users/{0}/permissions/{1}' -f $UserId,$Id)
 
-    return Invoke-CosmosDbRequest @PSBoundParameters `
+    $null = Invoke-CosmosDbRequest @PSBoundParameters `
         -Method 'Delete' `
         -ResourceType 'permissions' `
         -ResourcePath $resourcePath